+++ ping6-datasend/src/stream/packet.rs
@@ -24,7 +24,7 @@ pub fn make_stream_client_icmpv6_packet(
 ) -> Bytes {
     debug_assert!(!flags.test(StreamPacketFlags::WS));
 
-    let targlen = STREAM_CLIENT_FULL_HEADER_SIZE as usize + payload.len();
+    let targlen = STREAM_CLIENT_FULL_HEADER_SIZE + payload.len();
     let buflen = packet_buff.len();
     if buflen < targlen {
         packet_buff.reserve(targlen - buflen);
@@ -35,7 +35,7 @@ pub fn make_stream_client_icmpv6_packet(
         let mut packet = MutableIcmpv6Packet::new(packet_buff)
             .expect("buffer big enough for the payload");
         debug_assert!(packet.payload().len()
-            >= STREAM_CLIENT_HEADER_SIZE as usize + payload.len());
+            >= STREAM_CLIENT_HEADER_SIZE + payload.len());
 
         packet.set_icmpv6_type(Icmpv6Types::EchoRequest);
         packet.set_icmpv6_code(Icmpv6Codes::NoCode);
